Config, BasicConfig, MapConfigMapper, and YamlConfigMapper are all confusing names; Especially if another element in the API has a "Config" class, like many do.
Changes are made to provide backwards compatability with plugins still utilizing Yamler.
