+++ Yamler-Core/src/main/java/net/cubespace/Yamler/Config/ConfigBasic.java
@@ -1,0 +1,5 @@
+package net.cubespace.Yamler.Config;
+
+@Deprecated
+public class ConfigBasic extends BaseConfig {
+}
