+++ Yamler-Core/src/main/java/net/cubespace/Yamler/Config/YamlConfigMapper.java
@@ -1,4 +1,5 @@
 package net.cubespace.Yamler.Config;
 
+@Deprecated
 public class YamlConfigMapper extends BaseConfigMapper {
 }
